Add unit test for Annotation::renderToImage
This test checks the expected ratio of pixels in known color inside the
rendered image, with minor tolerance, in order to improve test
maintainability.

Note: The test covers ArthurOutputDev, which requires a QGuiApplication
due to internal usage of QRawFont. Therefore we have a dependency
on running X during test execution.