Use glyph widths from FreeType rather than what the file says
pdf files contain lists of glyph widths even for fonts that are
not embedded.  This can lead to problems when the exact font does
not exist on the system, and FreeType substitutes another similar
one.  In that case poppler currently uses the glyph widths from
the pdf file, which however may not match the font that is
actually used.