FlateStream: don't eat Adler-32 checksum
The FlateStream::getHuffmanCodeWord function always reads 'maxLen' bits
from the stream, even if it has already read enough bits to find a match
in the code tables. It may eat an extra byte from past the end of the
compressed data, where the Adler-32 checksum is. This will make
verification fail if we decide to verify the checksum.

Keep the Adler-32 checksum unread by changing the getHuffmanCodeWord
function to stop reading when it has read enough bits to find a match.