FlateStream: Verify Adler-32 checksum
A variant of the PDFex CBC gadget attack modifies Deflate compressed
data to assist with the creation of CBC gadgets. The attacker can use
CBC gadgets to prefix a URL to an encrypted plaintext, to create a
hyperlink which will leak decrypted data when clicked. However,
modifying the deflate data results in several random bytes in the
decompressed data, so we can limit this attack by verifying the Deflate
data's Adler-32 checksum.

See https://pdf-insecurity.org/encryption/encryption.html

Issue #814