Splash: Enhance bilinear scaling for downsampling
This is an additional attempt to counter #950.

Splash bilinear scaling was originally limited to up sampling. With small
enhancements we can now use it for down sampling too.

It is useful because the current Splash down sampling algorithm
(weighted average) is notably blurry when scaling down only a bit.
Bilinear is an improvement there, so we use it above 75%.

On the other hand it should never be used for down sampling below 50%,
because bilinear will then sample non-adjacent pixels.