CMS: rebuild ICC transforms when the display profile changes
The latest additions to the transparency group handling mean that the
GfxState's display ICC profile can change in the middle of rendering.
As a consequence all prebuilt and cached ICC transforms need to be
either rebuilt (as happens for Default{Gray,RGB,CMYK}ColorSpace} or
need to invalidated (as happens for the ICC profile cache).