Convert st_mtim reference to returning a copy
The reason why we need to do this is because on AIX, st_mtim is not
defined as a timespec, but rather, as a a struct { time_t, int },
which isn't compatible. As such, we can't return a direct reference,
nor just pass through the struct itself (except the macOS case), so
we instead return a new struct with the same field values.

I think the performance impact should be minimal.