Fix timeout in InterruptibleQueue.join
Previously, the 'join' method of InterruptibleQueue
realized 'no timeout' by specifying a very large
timeout parameter, viz. sys.float_info.max.

Supposedly this used to work, but with current Python
(e.g., the Python 3.10.7 on today's Debian testing)
this throws an error:

   OverflowError: timestamp too large to convert to C _PyTime_t

On the other hand, according to

  https://docs.python.org/3/library/threading.html#threading.Condition.wait

I can simply omit the method argument to also disable
the timeout.  This is what this patch does.