Remove SplashICCTransform altogether from drawImage functionality
The whole purpose of SplashICCTransform is to draw images with an ICC
colorspace. Unfortunately, this just doubles functionality from
GfxImageColorMap which can already call get{RGB,CMYK,...}Line of the
respective GfxICCBasedColorSpace, which does the color transform anyway.

Note that this commutes the order in which the image is scaled and
transformed in Splash::drawImage. So far, it is first scaled then the
ICC transform applied. Now, it will be the other way around, first color
transformed then scaled. This can potentially have speed implications, as
well as changes the color space in which the bilinear interpolation is
done for upscaling.